Changes to the signal processing code (#20)
* Added comments, error catch for incorrect usb port, and changed constants to class variables

* Changed ir durations to variables

* Fixed serial backlog issue. Refactored Message class and changed interactions with it.

* Added debug mode to obstacle avoidance loop

* Changed read_message from a loop to a single fire buffer scrape

* Fix minor bugs

* Fix big bugs

I changed the delimiter from '\n' to '|' since '\n's were getting lost.
Also made sure that we do not lose messages if we read while a message
is being received and thus not get the entire message in one go. The
partial message is stored and used during the next call to
read_message().