Baudrate (#45)
* Update user device for higher baudrate

* changed baud rate and fixed read bug with asimple delay

* Change EV3 baudrate

* Adjust for radio delay and IR ordering

* Merge master into baudrate

* Adjust delays
HC sensors have slightly different timeouts so needed to increase
minimum loop delay to the longest timeout. There is plenty of spare time
so we can either send multiple IR readings every loop or do the PID
calculations on Arduino rather than EV3.

* Ignore invalid characters from serial
We were getting the exceptions again so fixed it like this. We should
probably investigate what was causing it though.

* Adjust radio delay again